Move all metadata related scripts to a metadata directory
